Prevent Imagemagick to generate greyscale 8bit PNG
Imagemagick has a smart way to encode PNG, and switch to greyscale automagically if only grey pixels are detected. This is nice for web, for saving file weight, but for Pepper&Carrot ; it scrumbles the profile of the pages generated as PNG when the renderfarm hit a page with pencil/greyscale content.
Ref: http://www.imagemagick.org/Usage/formats/#png_write  ,  -define png:color-type=2