new feature: txt-only , bump to 6.0
Export a new subfolder in "hi-res" named "txt-only" with a rendering of the SVG without a background ( transparent ) as PNG. This files are handy in multiple case: for future project of book-publishing ; or for server-side less computer intensive rendering of layered picture, or for the creation in local of the comic page ( to link a dynamic layer, and get feedback of the txt ). 
Generating this pass has a big impact on performances, but the result worth it.