Push online the 'lang' subfolder with all SVGs.
This affect the performance of LFTP (so many more files to scan) but this system also offer hosting of the source SVG files in an alternative place than Github. Sources SVG files should be easy to reach, in case of proposing a downloader of global sources. I keep excluding the *.git subfolder for now ; impact on LFTP was too high.